CSU Coding Competition
One of the creators of this event, Max, looooooooves pizza. He loves it so much, that he is willing to buy pizzas for all of his friends ❤️. However, Max doesn't have a never ending bank account , and so he can't always afford to buy all of the pizza he wants.

Max's favorite pizza restaurant is Krazy Karls. Given a list of Krazy Karls famous pizzas, each represented as the cost of that pizza, Max wants us to find the subset of pizzas with the highest cost that is within his budget. (He likes to spoil his friends...)

This problem is brought to you by CSU Professor Wim Bohm!